requirements.txt
-r services/requirements.txt

faster-whisper>=1.0.0
yt-dlp>=2025.1.15
ffmpeg-python>=0.2.0
numpy>=1.26.0
scipy>=1.12.0
websockets>=12.0
rich>=13.7.0
httpx>=0.27.0
